requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.3.0
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.1
frozenlist==1.6.2
fsspec==2025.5.1
hydra-core==1.3.2
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.8
lightning-utilities==0.14.3
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.4.4
networkx==3.4.2
nibabel==5.3.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1.post0
pytz==2025.2
PyYAML==6.0.2
seaborn==0.13.2
six==1.17.0
sympy==1.14.0
torch==2.7.0
torchaudio==2.7.0
torchmetrics==1.7.2
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
typing_extensions==4.14.0
tzdata==2025.2
yarl==1.20.0
